+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 * Configures CI to run on pull request.
 
+* Add `Kaffe.MessageHander` behaviour. To utilize it, add the behaviour to your configured 
+`message_hander` and `@impl Kaffe.MessageHander` on `handle_messages/1`.
+
 ### Fixes
 
 * Stops compiler warnings on duplicate doc definitions